# grpc-gateway/dto_message_service.go
// This file generated automatically by gRPC gateway generator
// Generated at: 2024-04-09 12:09:49

package main

import messagepb "github.com/espitman/gws-chat/pkg/protos/protogen/message-service"

type messageServiceV1AddMemberToRoomResponseDto struct {
	ResponseDto
	Payload messagepb.V1AddMemberToRoomResponse
}

type messageServiceV1AddMessageResponseDto struct {
	ResponseDto
	Payload messagepb.V1AddMessageResponse
}

type messageServiceV1CreateRoomResponseDto struct {
	ResponseDto
	Payload messagepb.V1CreateRoomResponse
}

type messageServiceV1GetRoomResponseDto struct {
	ResponseDto
	Payload messagepb.V1GetRoomResponse
}

type messageServiceV1GetRoomMessagesResponseDto struct {
	ResponseDto
	Payload messagepb.V1GetRoomMessagesResponse
}