# grpc-gateway/dto_user_service.go
// This file generated automatically by gRPC gateway generator
// Generated at: 2024-04-09 12:09:49

package main

import userpb "github.com/espitman/gws-chat/pkg/protos/protogen/user-service"

type userServiceV1GetResponseDto struct {
	ResponseDto
	Payload userpb.V1GetResponse
}

type userServiceV1GetAllResponseDto struct {
	ResponseDto
	Payload userpb.V1GetAllResponse
}

type userServiceV1LoginResponseDto struct {
	ResponseDto
	Payload userpb.V1LoginResponse
}